requirements.txt
blinker==1.9.0
cffi==2.0.0
click==8.1.8
cryptography==41.0.7
Flask==3.0.0
Flask-Cors==4.0.0
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pexpect==4.9.0
ptyprocess==0.7.0
pycparser==2.23
python-dotenv==1.0.0
Werkzeug==3.1.5
zipp==3.23.0
